init modules via fx
