+++ controller/BUILD.bazel
@@ -3,6 +3,7 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "controller",
     srcs = [
+        "comparer.go",
         "controller.go",
         "distance_filter.go",
         "errors.go",
