+++ controller/gettargetgraph_test.go
@@ -46,7 +46,7 @@ func TestGetTargetGraph_CacheMiss_NoSend(t *testing.T) {
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).
 			Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte{})}, nil),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -71,7 +71,7 @@ func TestGetTargetGraph_StorageError_Propagates(t *testing.T) {
 	stream.EXPECT().Context().Return(context.Background())
 	storagemock := storagemock.NewMockStorage(ctrl)
 	storagemock.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{}, expected)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: storagemock,
 	})
@@ -97,7 +97,7 @@ func TestGetTargetGraph_DecodeError_ReturnsError(t *testing.T) {
 		storagemock.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("treehash-abc"))}, nil),
 		storagemock.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("bad-bytes"))}, nil),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: storagemock,
 	})
@@ -130,7 +130,7 @@ func TestGetTargetGraph_SendsWhenItemPresent(t *testing.T) {
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("treehash-xyz"))}, nil),
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser(buf.Bytes())}, nil),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -152,7 +152,7 @@ func TestGetTargetGraph_BuildDescriptionMissingRequiredFields_ReturnsError(t *te
 	stream := tangomock.NewMockTangoServiceGetTargetGraphYARPCServer(ctrl)
 	stream.EXPECT().Context().Return(context.Background())
 	store := storagemock.NewMockStorage(ctrl)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -173,7 +173,7 @@ func TestGetTargetGraph_MissingBuildDescription_ReturnsError(t *testing.T) {
 	stream := tangomock.NewMockTangoServiceGetTargetGraphYARPCServer(ctrl)
 	stream.EXPECT().Context().Return(context.Background())
 	store := storagemock.NewMockStorage(ctrl)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -202,7 +202,7 @@ func TestGetTargetGraph_TreehashNotFound_NoError(t *testing.T) {
 	graphReader.EXPECT().Read().Return(nil, io.EOF).Times(1)
 	graphReader.EXPECT().Close().Return(nil)
 	orchestrator.EXPECT().GetTargetGraph(gomock.Any(), gomock.Any()).Return(graphReader, nil)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:       zaptest.NewLogger(t),
 		Storage:      store,
 		Orchestrator: orchestrator,
@@ -220,7 +220,7 @@ func TestGetTargetGraph_TreehashReadError(t *testing.T) {
 	stream.EXPECT().Context().Return(context.Background())
 	store := storagemock.NewMockStorage(ctrl)
 	store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: &errReadCloser{err: errors.New("readfail")}}, nil)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -240,7 +240,7 @@ func TestGetTargetGraph_GraphFetchError(t *testing.T) {
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("treehash-abc"))}, nil),
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{}, errors.New("graph error")),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -264,7 +264,7 @@ func TestGetTargetGraph_GraphReadError(t *testing.T) {
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("treehash-abc"))}, nil),
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: &errReadCloser{err: errors.New("readfail")}}, nil),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -292,7 +292,7 @@ func TestGetTargetGraph_StreamSendError(t *testing.T) {
 		storagemock.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("treehash-abc"))}, nil),
 		storagemock.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser(buf.Bytes())}, nil),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: storagemock,
 	})
@@ -321,7 +321,7 @@ func TestGetTargetGraph_GraphNotFound_FallsThrough(t *testing.T) {
 	graphReader.EXPECT().Read().Return(nil, io.EOF).Times(1)
 	graphReader.EXPECT().Close().Return(nil)
 	orch.EXPECT().GetTargetGraph(gomock.Any(), gomock.Any()).Return(graphReader, nil)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:       zaptest.NewLogger(t),
 		Storage:      store,
 		Orchestrator: orch,
@@ -343,7 +343,7 @@ func TestGetTargetGraph_GraphReadCancelled(t *testing.T) {
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{ReadCloser: newMockReadCloser([]byte("treehash-abc"))}, nil),
 		store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{}, errors.New("context canceled")),
 	)
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:  zaptest.NewLogger(t),
 		Storage: store,
 	})
@@ -367,7 +367,7 @@ func TestGetTargetGraph_OrchestratorCancelled(t *testing.T) {
 	store.EXPECT().Get(gomock.Any(), gomock.Any()).Return(storage.DownloadResponse{}, &storage.NotFoundError{Path: "x"})
 	orch := orchestratormock.NewMockOrchestrator(ctrl)
 	orch.EXPECT().GetTargetGraph(gomock.Any(), gomock.Any()).Return(nil, errors.New("context canceled"))
-	c := NewController(context.Background(), Params{
+	c := NewController(Params{AppCtx: context.Background(), 
 		Logger:       zaptest.NewLogger(t),
 		Storage:      store,
 		Orchestrator: orch,
