+++ example/BUILD.bazel
@@ -7,13 +7,12 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//config",
-        "//controller",
         "//core/git",
-        "//core/repomanager",
         "//core/storage",
         "//core/storage/disk",
-        "//orchestrator",
+        "//module",
         "//tangopb",
+        "@org_uber_go_fx//:fx",
         "@org_uber_go_yarpc//:yarpc",
         "@org_uber_go_yarpc//api/transport",
         "@org_uber_go_yarpc//transport/grpc",
